requirements.txt
pdf2image
opencv-python
numpy
pandas
pytesseract
streamlit
requests
Pillow
camelot-py[cv]
openpyxl